Version 25.0, fix windows compiles

+++ src/hud.cpp
@@ -534,11 +534,11 @@ void DrawLuaLogs()
 
 void GetTimeString(int currtime, char *string, int length)
 {
-	int years, days, hours, minutes, seconds, milliseconds;
+	long long years, days, hours, minutes, seconds, milliseconds;
 	if (length == 0)
 	{
-		years = currtime/31557600000;
-		currtime = currtime%31557600000;
+		years = currtime/31557600000ULL;
+		currtime = currtime%31557600000ULL;
 		days = currtime/86400000;
 		currtime = currtime%86400000;
 	}
